Increase FileInfo's size from int to bigint

+++ osf/models/fileinfo.py
@@ -5,4 +5,4 @@ from osf.models.base import BaseModel, ObjectIDMixin
 class FileInfo(ObjectIDMixin, BaseModel):
     """Saves extra information about a file."""
     file = models.OneToOneField('osf.BaseFileNode', on_delete=models.CASCADE)
-    file_size = models.IntegerField()
+    file_size = models.BigIntegerField()
